+++ client/lib/js/plugins/bigchaindb_ledgerplugin.js
